# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ShareHub API
  version: 0.2.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/notes:
    get:
      summary: 笔记分页列表
      description: 需要当前用户身份；当前实现只返回请求用户自己的笔记；`page`、`pageSize` 传入小于 1 的值时都会按 `1` 兜底；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`；`status` 传空字符串或仅空白时按未传处理，不额外收窄结果；带前后空白时会先裁剪再参与筛选
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            default: 10
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotePageResponse'
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: 创建笔记
      description: 需要当前用户身份；创建后笔记归属当前请求用户；`status` 未传、为 `null`、空字符串或仅空白时默认保存为 `DRAFT`，带前后空白时会先裁剪再保存；`visibility` 为 `null`、空字符串或仅空白时按未传处理，最终存为 `null`；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteUpsertRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteResponse'
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/notes/{id}:
    get:
      summary: 获取笔记详情
      description: 公开已发布笔记支持匿名或其他登录用户直接读取；当前用户访问自己的笔记时仍可读取草稿与非公开内容；若已带用户身份且用户被封禁返回 `403 USER_BANNED`，其余不可访问场景统一按不存在处理
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteResponse'
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 笔记不存在（错误码 `NOTE_NOT_FOUND`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: 更新笔记
      description: 需要当前用户身份；当前实现只允许更新自己的笔记；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`，访问他人笔记时按不存在处理；`status` 传 `null`、空字符串或仅空白时按未传处理，保留原状态；带前后空白时会先裁剪再保存；`visibility` 传 `null`、空字符串或仅空白时按未传处理，最终存为 `null`
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteUpsertRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteResponse'
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 笔记不存在（错误码 `NOTE_NOT_FOUND`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: 删除笔记
      description: 需要当前用户身份；当前实现只允许删除自己的笔记；未带身份时返回 `401 NOT_LOGGED_IN`，用户被封禁时返回 `403 USER_BANNED`，访问他人笔记时按不存在处理
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '401':
          description: 未登录（错误码 `NOT_LOGGED_IN`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 用户被封禁（错误码 `USER_BANNED`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 笔记不存在（错误码 `NOTE_NOT_FOUND`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        code:
          type: string
          description: 业务错误码，便于前端区分类型
        message:
          type: string
          description: 出错提示
...
    NoteDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        contentMd:
          type: string
        visibility:
          type: string
          nullable: true
          description: 可见性；传空字符串或仅空白时最终存为 null
        status:
          type: string
          nullable: true
          description: 笔记状态；创建时未传或空白会回落为 DRAFT
      required:
        - id
        - title
        - contentMd
    NoteUpsertRequest:
      type: object
      properties:
        title:
          type: string
        contentMd:
          type: string
        visibility:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
      required:
        - title
        - contentMd
    NoteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/NoteDto'
        message:
          type: string
    NotePageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            total:
              type: integer
              format: int64
            page:
              type: integer
            pageSize:
              type: integer
              items:
                type: array
                items:
                  $ref: '#/components/schemas/NoteDto'
        message:
          type: string
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: string
          example: DELETED
        message:
          type: string
...